requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
colored==1.4.2
idna==2.9
ImageHash==4.1.0
imutils==0.5.3
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pytesseract==0.3.4
PyWavelets==1.1.1
requests==2.23.0
scipy==1.4.1
six==1.15.0
urllib3==1.25.9
